goodreads/requirements.txt
BeautifulSoup4
python-dateutil
requests
